requirements.txt
feedparser
requests
beautifulsoup4
transformers
huggingface_hub
torch
python-docx